(Save / Get) Sessions

+++ School.API/DTOs/Common/ResponseDto.cs
@@ -2,7 +2,7 @@
 {
     public class ResponseDto
     {
-        public int VID { get; set; }
+        public decimal VID { get; set; }
         public int ReturnCode { get; set; }
         public string ReturnMessage { get; set; }
     }
